npcs

+++ app/main.scala
@@ -28,20 +28,20 @@ object main extends App with DataFunc{
       if (adress(0).equals("TCP_IP")) {
         val addr = adress(1)
         val arr = addr.split("\\.")
-        val a1 = Integer.valueOf(arr(0)).byteValue()
-        val a2 = Integer.valueOf(arr(1)).byteValue()
-        val a3 = Integer.valueOf(arr(2)).byteValue()
-        val a4 = Integer.valueOf(arr(3)).byteValue()
+        val a1 = Integer.valueOf(arr(0)).shortValue()
+        val a2 = Integer.valueOf(arr(1)).shortValue()
+        val a3 = Integer.valueOf(arr(2)).shortValue()
+        val a4 = Integer.valueOf(arr(3)).shortValue()
         val port = Integer.valueOf(arr(4)).shortValue()
         TCP_IP = (a1, a2, a3, a4, port)
       }
       if (adress(0).equals("MAP_IP")) {
         val addr = adress(1)
         val arr = addr.split("\\.")
-        val a1 = Integer.valueOf(arr(0)).byteValue()
-        val a2 = Integer.valueOf(arr(1)).byteValue()
-        val a3 = Integer.valueOf(arr(2)).byteValue()
-        val a4 = Integer.valueOf(arr(3)).byteValue()
+        val a1 = Integer.valueOf(arr(0)).shortValue()
+        val a2 = Integer.valueOf(arr(1)).shortValue()
+        val a3 = Integer.valueOf(arr(2)).shortValue()
+        val a4 = Integer.valueOf(arr(3)).shortValue()
         val port = Integer.valueOf(arr(4)).shortValue()
         MAP_IP = (a1, a2, a3, a4, port)
       }
@@ -69,8 +69,8 @@ object Settings{
   var nmSQL:String = _
   var dbName:String = _
 
-  var TCP_IP:(Byte,Byte,Byte,Byte,Short) = _
-  var MAP_IP:(Byte,Byte,Byte,Byte,Short) = _
+  var TCP_IP:(Short,Short,Short,Short,Short) = _
+  var MAP_IP:(Short,Short,Short,Short,Short) = _
   lazy val hostPath: String = TCP_IP._1 + "." + TCP_IP._2 + "." + TCP_IP._3 + "." + TCP_IP._4
   lazy val mapHostPath: String = MAP_IP._1 + "." + MAP_IP._2 + "." + MAP_IP._3 + "." + MAP_IP._4
   def ActorPath(path:String) : String = "akka://ActorSystem/user/"+path
